middleware su tutti i Controller; possibilità di vedere il sito anche da non loggati.

+++ app/Http/Controllers/UserController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Event;
 use App\Group;
+use App\Itinerary;
 use App\User;
 use App\UserGroup;
 use Illuminate\Http\Request;
@@ -30,20 +32,36 @@ class UserController extends Controller
      */
     public function index()
     {
-       /* $user = DB::table('users')
-            ->leftJoin('users_groups', 'users.id', '=', 'users_groups.group_id')
-            ->select('users.id' ,'users.name', 'users.email', 'users_groups.group_id')
-            ->get();*/
-       $user = User::paginate(10);
+
+        $permission = $this->authentication();
+
+        if($permission){
+
+            /* $user = DB::table('users')
+           ->leftJoin('users_groups', 'users.id', '=', 'users_groups.group_id')
+           ->select('users.id' ,'users.name', 'users.email', 'users_groups.group_id')
+           ->get();*/
+            $user = User::paginate(10);
 
 
-        /* $user_group = DB::table('users_groups')
-             ->where('users_groups.user_id', '=', 1)
-             ->select('users_groups.group_id')
-             ->first();
-         */
+            /* $user_group = DB::table('users_groups')
+                 ->where('users_groups.user_id', '=', 1)
+                 ->select('users_groups.group_id')
+                 ->first();
+             */
 
-        return view('admin/user/index', ['user' => $user]);
+            return view('admin/user/index', ['user' => $user]);
+
+        }else{
+
+            $itineraries = Itinerary::take(4)->get();
+            $events = Event::take(3)->get();
+            return \Illuminate\Support\Facades\View::make('welcome')
+                ->with('permission', $permission )
+                ->with('itineraries', $itineraries)
+                ->with('events', $events);
+        }
+
     }
 
 
@@ -56,11 +74,27 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        $user = DB::table('users')
-            ->where('id', $id)
-            ->first();
 
-        return view('admin/user/edit', ['user' => $user]);
+        $permission = $this->authentication();
+
+        if($permission){
+
+            $user = DB::table('users')
+                ->where('id', $id)
+                ->first();
+
+            return view('admin/user/edit', ['user' => $user]);
+
+        }else{
+
+            $itineraries = Itinerary::take(4)->get();
+            $events = Event::take(3)->get();
+            return \Illuminate\Support\Facades\View::make('welcome')
+                ->with('permission', $permission )
+                ->with('itineraries', $itineraries)
+                ->with('events', $events);
+        }
+
     }
 
     /**
@@ -72,18 +106,34 @@ class UserController extends Controller
      */
     public function store(Request $request, $id)
     {
-        $this->validateItems($request);
 
-        DB::table('users')
-            ->where('id', $id)
-            ->update([
-                'name'          => $request['name'],
-                'email'         => $request['email'],
-            ]);
+        $permission = $this->authentication();
 
-        flash('Success')->success();
+        if($permission){
 
-        return redirect('admin/user/index');
+            $this->validateItems($request);
+
+            DB::table('users')
+                ->where('id', $id)
+                ->update([
+                    'name'          => $request['name'],
+                    'email'         => $request['email'],
+                ]);
+
+            flash('Success')->success();
+
+            return redirect('admin/user/index');
+
+        }else{
+
+            $itineraries = Itinerary::take(4)->get();
+            $events = Event::take(3)->get();
+            return \Illuminate\Support\Facades\View::make('welcome')
+                ->with('permission', $permission )
+                ->with('itineraries', $itineraries)
+                ->with('events', $events);
+        }
+
     }
 
     /**
@@ -96,7 +146,12 @@ class UserController extends Controller
 
     public function delete($id)
     {
-        $user = User::find($id);
+
+        $permission = $this->authentication();
+
+        if($permission){
+
+            $user = User::find($id);
 
 
             $user->groupRel()->wherePivot('user_id','=',$id)->detach();
@@ -105,103 +160,137 @@ class UserController extends Controller
                 ->where('id', $id)
                 ->delete();
 
-        flash('Deleted')->error();
+            flash('Deleted')->error();
 
-        return redirect()->back();
+            return redirect()->back();
+
+        }else{
+
+            $itineraries = Itinerary::take(4)->get();
+            $events = Event::take(3)->get();
+            return \Illuminate\Support\Facades\View::make('welcome')
+                ->with('permission', $permission )
+                ->with('itineraries', $itineraries)
+                ->with('events', $events);
+        }
+
     }
 
 
     //funzione per assegnare gruppi
     public function showAssignment($id){
 
-        /*$user = DB::table('users')
+
+        $permission = $this->authentication();
+
+        if($permission){
+
+            /*$user = DB::table('users')
             ->where('id', $id)
             ->first();*/
 
-        $group = DB::table('groups')
-            ->get();
+            $group = DB::table('groups')
+                ->get();
 
 
 
-        $user = User::find($id);
+            $user = User::find($id);
 
 
-        //prendi tutti i gruppi per cui l'utente non ha relazione nella users_groups
+            //prendi tutti i gruppi per cui l'utente non ha relazione nella users_groups
 
-        //conto quanti users_group per user = $id ci sono
-        $countUsersGroups = User::find($id)->groupRel()->wherePivot('user_id', '=', $id)->count();
-        $var = User::find($id)->groupRel()->wherePivot('user_id', '=', $id)->get();
+            //conto quanti users_group per user = $id ci sono
+            $countUsersGroups = User::find($id)->groupRel()->wherePivot('user_id', '=', $id)->count();
+            $var = User::find($id)->groupRel()->wherePivot('user_id', '=', $id)->get();
 
-        //per ogni $var prendi i gruppi assegnati poi
-        // $gruppi = Group::find(1)->userRel()->wherePivot('user_id', '=', $id)->where('id', '!=', $var->group_id);
+            //per ogni $var prendi i gruppi assegnati poi
+            // $gruppi = Group::find(1)->userRel()->wherePivot('user_id', '=', $id)->where('id', '!=', $var->group_id);
 
-        //echo $gruppi;
-        //echo $var;
-
-
-
-        //conto i gruppi
-        //$countGroup = Group::all()->count();
+            //echo $gruppi;
+            //echo $var;
 
 
-        return view('admin/user/assign', ['user' => $user], ['group' => $group]);
+
+            //conto i gruppi
+            //$countGroup = Group::all()->count();
+
+
+            return view('admin/user/assign', ['user' => $user], ['group' => $group]);
+        }else{
+
+            $itineraries = Itinerary::take(4)->get();
+            $events = Event::take(3)->get();
+            return \Illuminate\Support\Facades\View::make('welcome')
+                ->with('permission', $permission )
+                ->with('itineraries', $itineraries)
+                ->with('events', $events);
+        }
+
+
 
     }
 
     public function saveAssignment(Request $request){
 
-       /* DB::table('users_groups')
-            ->insert([
-                'user_id'          => $request['user_id'],
-                'group_id'         => $request['group_id'],
-            ]);*/
+        $permission = $this->authentication();
+
+        if($permission){
+
+            $ngroup = Group::all()->count();
 
 
+            $user = User::find($request['user_id']);
 
+            $var= 0;
+            $var1 = $request['group_id'];
+            foreach ($user->groupRel as $role)
+            {
 
-        $ngroup = Group::all()->count();
+                if($role->pivot->group_id != $var1 ){
 
+                    $var = $var +1;
+                }else{
 
-        $user = User::find($request['user_id']);
+                    $var = $var - 1;
+                }
 
-        $var= 0;
-        $var1 = $request['group_id'];
-        foreach ($user->groupRel as $role)
-        {
-
-            if($role->pivot->group_id != $var1 ){
-
-                $var = $var +1;
-            }else{
-
-                $var = $var - 1;
             }
 
+
+            if($var < $ngroup && $var !== 0 && $var >= 1){
+
+                DB::table('users_groups')
+                    ->insert([
+                        'user_id'          => $request['user_id'],
+                        'group_id'         => $request['group_id'],
+                    ]);
+
+            }
+
+            //se label assegned non c'è allora inserisce
+            if(!isset( $_GET['group'])){
+                DB::table('users_groups')
+                    ->insert([
+                        'user_id'          => $request['user_id'],
+                        'group_id'         => $request['group_id'],
+                    ]);
+
+            }
+
+
+            return redirect('admin/user/index');
+
+
+        }else{
+
+            $itineraries = Itinerary::take(4)->get();
+            $events = Event::take(3)->get();
+            return \Illuminate\Support\Facades\View::make('welcome')
+                ->with('permission', $permission )
+                ->with('itineraries', $itineraries)
+                ->with('events', $events);
         }
 
-
-        if($var < $ngroup && $var !== 0 && $var >= 1){
-
-            DB::table('users_groups')
-                ->insert([
-                    'user_id'          => $request['user_id'],
-                    'group_id'         => $request['group_id'],
-                ]);
-
-        }
-
-        //se label assegned non c'è allora inserisce
-        if(!isset( $_GET['group'])){
-            DB::table('users_groups')
-                ->insert([
-                    'user_id'          => $request['user_id'],
-                    'group_id'         => $request['group_id'],
-                ]);
-
-        }
-
-
-        return redirect('admin/user/index');
 
     }
 
@@ -209,10 +298,27 @@ class UserController extends Controller
     //funzione per profilo
     public function settings(){
 
-        $user = DB::table('users')
-            ->first();
+        $permission = $this->authentication();
 
-        return view('admin/settings', ['user' => $user]);
+        if($permission){
+
+            $user = DB::table('users')
+                ->first();
+
+            return view('admin/settings', ['user' => $user]);
+
+
+        }else{
+
+            $itineraries = Itinerary::take(4)->get();
+            $events = Event::take(3)->get();
+            return \Illuminate\Support\Facades\View::make('welcome')
+                ->with('permission', $permission )
+                ->with('itineraries', $itineraries)
+                ->with('events', $events);
+        }
+
+
     }
 
 
@@ -229,13 +335,60 @@ class UserController extends Controller
     public function removeAssignment($userId,$groupId)
     {
 
-        $user = User::find($userId);
+        $permission = $this->authentication();
 
-        $user->groupRel()->wherePivot('user_id','=',$userId)
-                        ->wherePivot('group_id' , '=', $groupId)
-                        ->detach();
+        if($permission){
 
-        return redirect()->back();
+            $user = User::find($userId);
+
+            $user->groupRel()->wherePivot('user_id','=',$userId)
+                ->wherePivot('group_id' , '=', $groupId)
+                ->detach();
+
+            return redirect()->back();
+
+        }else{
+
+            $itineraries = Itinerary::take(4)->get();
+            $events = Event::take(3)->get();
+            return \Illuminate\Support\Facades\View::make('welcome')
+                ->with('permission', $permission )
+                ->with('itineraries', $itineraries)
+                ->with('events', $events);
+        }
+
+    }
+
+
+
+    public function authentication(){
+
+        $permission = false;
+
+        if(Auth::check()){
+
+            $id = Auth::user()->id;
+
+            $user = User::find($id);
+
+            if(isset($user->groupRel)){
+
+                foreach ($user->groupRel as $item) {
+
+                    $group = Group::all()->where('id', $item->pivot->group_id)->first();
+
+                    if ($group->name == 'admin') {
+
+                        $permission = true;
+
+                    }
+                }
+
+            }
+        }
+
+        return $permission;
+
     }
